testing

+++ scripts/install-ruby.sh
@@ -11,8 +11,13 @@
 # source $HOME/.rvm/scripts/rvm
 # rvm install 1.8.7
 
-wget ftp://ftp.ruby-lang.org/pub/ruby/1.8/ruby-1.8.7-p248.tar.gz -O /tmp/ruby.tar.gz
-cd /tmp && tar -zxvf ruby.tar.gz
-cd ruby*
-./configure && make && make install
-cd .. && rm -rf ruby*
+if ruby -v
+then
+    echo "Ruby already installed"
+else
+    wget ftp://ftp.ruby-lang.org/pub/ruby/1.8/ruby-1.8.7-p248.tar.gz -O /tmp/ruby.tar.gz
+    cd /tmp && tar -zxvf ruby.tar.gz
+    cd ruby*
+    ./configure && make && make install
+    cd .. && rm -rf ruby*
+fi
